Split out fixtures to fixtures.ts
