+++ test/UpdateLimitOrder.ts
@@ -1,88 +1,6 @@
 import { expect } from 'chai'
-import hre, { ethers } from 'hardhat'
 import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
-
-export async function get_setup_values() {
-  const [owner, acc1, acc2] = await ethers.getSigners();
-
-  // Deploy heap libraries
-  const max = await ethers.getContractFactory("MaxLinkedListLib");
-  const maxList = await max.deploy();
-  await maxList.deployed();
-
-  const min = await ethers.getContractFactory("MinLinkedListLib");
-  const minList = await min.deploy();
-  await minList.deployed();
-
-  // Deploy factory
-  const Factory = await ethers.getContractFactory("Factory", {
-    libraries: {
-      MaxLinkedListLib: maxList.address,
-      MinLinkedListLib: minList.address,
-    },
-  });
-  const factory = await Factory.deploy(owner.address);
-  await factory.deployed();
-
-  // Deploy router
-  const routerFactory = await ethers.getContractFactory("Router");
-  const router = await routerFactory.deploy(factory.address);
-  await router.deployed();
-
-  await factory.setRouter(router.address);
-
-  const token0_factory = await ethers.getContractFactory("TestERC20");
-  let token0 = await token0_factory.deploy("Test Token 0", "TEST 0");
-  await token0.deployed();
-
-  const token1_factory = await ethers.getContractFactory("TestERC20");
-  let token1 = await token1_factory.deploy("Test Token 1", "TEST 1");
-  await token1.deployed();
-
-  // Create the order book
-  const sizeTick = 100; // decimal=3 so multiples of 0.1
-  const priceTick = 10; // decimal=3 so multiples of 0.01
-  await factory.createOrderBook(token0.address, token1.address, 2, 1);
-
-  return {
-    factory,
-    router,
-    token0,
-    token1,
-    owner,
-    acc1,
-    acc2,
-    sizeTick,
-    priceTick,
-  };
-}
-
-export async function setup_and_deposit_in_vault_fixture() {
-  const { factory, router, token0, token1, owner, acc1, acc2, sizeTick, priceTick } =
-    await get_setup_values();
-
-  await token0.mint(acc1.getAddress(), "10000000000000");
-  await token0.connect(acc1).approve(router.address, "10000000000000");
-  await token0.mint(acc2.getAddress(), "10000000000000");
-  await token0.connect(acc2).approve(router.address, "10000000000000");
-
-  await token1.mint(acc1.getAddress(), "10000000000000");
-  await token1.connect(acc1).approve(router.address, "10000000000000");
-  await token1.mint(acc2.getAddress(), "10000000000000");
-  await token1.connect(acc2).approve(router.address, "10000000000000");
-
-  return {
-    factory,
-    router,
-    token0,
-    token1,
-    owner,
-    acc1,
-    acc2,
-    sizeTick,
-    priceTick,
-  };
-}
+import { setup_and_deposit_in_vault_fixture } from './helpers/fixtures';
 
 describe("Update limit order function", function () {
 
@@ -91,12 +9,12 @@ describe("Update limit order function", function () {
       setup_and_deposit_in_vault_fixture
     );
 
-    const startBalance = BigInt(await token0.balanceOf(acc1.address));
+    const startBalance = BigInt((await token0.balanceOf(acc1.address)).toString());
 
     // 2
-    await router.connect(acc1).createLimitOrder(0, 2, 1, 1, 0);
+    await router.connect(acc1).createLimitOrder(0, 2, 1, true, 0);
     // 3
-    await router.connect(acc1).createLimitOrder(0, 3, 1, 1, 0);
+    await router.connect(acc1).createLimitOrder(0, 3, 1, true, 0);
 
     // assert ids are correct
     const orders = await router.getLimitOrders(0);
@@ -129,9 +47,9 @@ describe("Update limit order function", function () {
     );
 
     // 2
-    await router.connect(acc1).createLimitOrder(0, 1, 10, 1, 0);
+    await router.connect(acc1).createLimitOrder(0, 1, 10, true, 0);
     // 3
-    await router.connect(acc1).createLimitOrder(0, 10, 1, 0, 0);
+    await router.connect(acc1).createLimitOrder(0, 10, 1, false, 0);
 
     await router.connect(acc1).updateLimitOrder(0, 2, 10, 1, 2);
 
@@ -145,12 +63,12 @@ describe("Update limit order function", function () {
     );
 
     // 2
-    await router.connect(acc1).createLimitOrder(0, 1, 1, 0, 0);
+    await router.connect(acc1).createLimitOrder(0, 1, 1, false, 0);
     // 3
-    await router.connect(acc1).createLimitOrder(0, 2, 2, 0, 0);
+    await router.connect(acc1).createLimitOrder(0, 2, 2, false, 0);
     // 4
 
-    await router.connect(acc1).createLimitOrder(0, 1, 3, 1, 0); // a0: 100, a1: 30
+    await router.connect(acc1).createLimitOrder(0, 1, 3, true, 0); // a0: 100, a1: 30
 
     const orders = await router.getLimitOrders(0);
     expect([4, 3, 2]).to.eql(orders[0]);
@@ -171,7 +89,7 @@ describe("Update limit order function", function () {
     );
 
     async function try_to_update() {
-      await router.connect(acc1).createLimitOrder(0, 7, 3, 0, 0);
+      await router.connect(acc1).createLimitOrder(0, 7, 3, false, 0);
       await router.connect(acc2).updateLimitOrder(0, 2, 1, 1, 0);
     }
 
@@ -185,7 +103,7 @@ describe("Update limit order function", function () {
       setup_and_deposit_in_vault_fixture
     );
 
-    await router.connect(acc1).createLimitOrder(0, 7, 3, 0, 0);
+    await router.connect(acc1).createLimitOrder(0, 7, 3, false, 0);
     await router.connect(acc1).cancelLimitOrder(0, 2);
     await router.connect(acc2).updateLimitOrder(0, 2, 1, 1, 0);
